deleted unnecessary eventlist component, changed seeds and deleted js seed data
